+++ FSICdata/ons/__init__.py
@@ -1,1 +1,1 @@
-from FSICdata.ons.ons import read_csv
+from FSICdata.ons.ons import get_csv, read_csv